@@ -3,11 +3,13 @@ using MediatR;
 using Microsoft.Extensions.DependencyInjection;
 using SFA.DAS.ApprenticeAan.Application.Regions.Queries.GetRegions;
 
-namespace SFA.DAS.ApprenticeAan.Application.Extensions
+namespace SFA.DAS.ApprenticeAan.Application.Extensions;
+
+[ExcludeFromCodeCoverage]
+public static class ServiceCollectionExtensions
 {
-    [ExcludeFromCodeCoverage]
-    public static class ServiceCollectionExtensions
+    public static void AddApplicationRegistrations(this IServiceCollection services)
     {
-        public static void AddApplicationRegistrations(this IServiceCollection services) { services.AddMediatR(typeof(GetRegionsQuery).Assembly); }
+        services.AddMediatR(typeof(GetRegionsQuery).Assembly);
     }
 }
