@@ -1,7 +1,7 @@
-﻿using SFA.DAS.Api.Common.AppStart;
+﻿using System.Diagnostics.CodeAnalysis;
+using SFA.DAS.Api.Common.AppStart;
 using SFA.DAS.Api.Common.Configuration;
-using SFA.DAS.SharedOuterApi.AppStart;
-using System.Diagnostics.CodeAnalysis;
+
 
 namespace SFA.DAS.ApprenticeAan.Api.AppStart;
 
