@@ -1,9 +1,10 @@
 ﻿using System.Diagnostics.CodeAnalysis;
 
-namespace SFA.DAS.ApprenticeAan.Application.Configuration;
+
+namespace SFA.DAS.ApprenticeAan.Infrastructure;
 
 [ExcludeFromCodeCoverage]
-public class AanHubApiConfiguration
+public class InnerApiConfiguration
 {
     public string Url { get; set; } = null!;
     public string Identifier { get; set; } = null!;
