@@ -4,35 +4,40 @@ using Microsoft.Extensions.Options;
 using RestEase.HttpClientFactory;
 using SFA.DAS.Api.Common.Infrastructure;
 using SFA.DAS.Api.Common.Interfaces;
-using SFA.DAS.ApprenticeAan.Api.Configuration;
 using SFA.DAS.ApprenticeAan.Api.HealthCheck;
+using SFA.DAS.ApprenticeAan.Application.Configuration;
 using SFA.DAS.ApprenticeAan.Application.Extensions;
 using SFA.DAS.ApprenticeAan.Application.Infrastructure;
-using SFA.DAS.ApprenticeAan.Application.Services;
+using SFA.DAS.ApprenticeAan.Configuration;
 using SFA.DAS.ApprenticeAan.Infrastructure;
-using SFA.DAS.SharedOuterApi.Configuration;
-using SFA.DAS.SharedOuterApi.Infrastructure;
-using SFA.DAS.SharedOuterApi.Infrastructure.HealthCheck;
-using SFA.DAS.SharedOuterApi.Interfaces;
-using SFA.DAS.SharedOuterApi.Services;
+using ApprenticeAccountsApiHealthCheck = SFA.DAS.ApprenticeAan.Api.HealthCheck.ApprenticeAccountsApiHealthCheck;
+using CoursesApiHealthCheck = SFA.DAS.ApprenticeAan.Api.HealthCheck.CoursesApiHealthCheck;
 
 namespace SFA.DAS.ApprenticeAan.Api.AppStart;
 
 [ExcludeFromCodeCoverage]
 public static class ServiceCollectionExtensions
 {
+    private static readonly string Ready = "ready";
     public static IServiceCollection AddServiceHealthChecks(this IServiceCollection services)
     {
         services.AddHealthChecks()
-            .AddCheck<ApprenticeAanInnerApiHealthCheck>(ApprenticeAanInnerApiHealthCheck.HealthCheckResultDescription,
+            .AddCheck<AanHubApiHealthCheck>(AanHubApiHealthCheck.HealthCheckResultDescription,
                 failureStatus: HealthStatus.Unhealthy,
-                tags: new[] { "ready" })
+                tags: new[] { Ready })
+            .AddCheck<CommitmentsV2InnerApiHealthCheck>(CommitmentsV2InnerApiHealthCheck.HealthCheckResultDescription,
+                failureStatus: HealthStatus.Unhealthy,
+                tags: new[] { Ready })
+            .AddCheck<ApprenticeAccountsApiHealthCheck>(ApprenticeAccountsApiHealthCheck.HealthCheckResultDescription,
+                failureStatus: HealthStatus.Unhealthy,
+                tags: new[] { Ready })
             .AddCheck<CoursesApiHealthCheck>(CoursesApiHealthCheck.HealthCheckResultDescription,
                 failureStatus: HealthStatus.Unhealthy,
-                tags: new[] { "ready" })
+                tags: new[] { Ready })
             .AddCheck<LocationsApiHealthCheck>(LocationsApiHealthCheck.HealthCheckResultDescription,
                 failureStatus: HealthStatus.Unhealthy,
-                tags: new[] { "ready" });
+                tags: new[] { Ready });
+        ;
         return services;
     }
 
@@ -41,42 +46,57 @@ public static class ServiceCollectionExtensions
         AddConfigurationOptions(services, configuration);
         services.AddHttpClient();
         services.AddTransient<IAzureClientCredentialHelper, AzureClientCredentialHelper>();
-        services.AddTransient(typeof(IInternalApiClient<>), typeof(InternalApiClient<>));
-        services.AddTransient<IAanHubApiClient<AanHubApiConfiguration>, AanHubApiClient>();
-        services.AddTransient<ILocationApiClient<LocationApiConfiguration>, LocationApiClient>();
-        services.AddTransient<IApprenticeAccountsApiClient<ApprenticeAccountsApiConfiguration>, ApprenticeAccountsApiClient>();
-        services.AddTransient<ICoursesApiClient<CoursesApiConfiguration>, CourseApiClient>();
+
         services.AddApplicationRegistrations();
         AddAanHubApiClient(services, configuration);
         AddCommitmentsV2ApiClient(services, configuration);
+        AddApprenticeAccountsApiClient(services, configuration);
+        AddCoursesApiClient(services, configuration);
+        AddLocationApiClient(services, configuration);
         return services;
     }
 
     private static void AddAanHubApiClient(IServiceCollection services, IConfiguration configuration)
     {
-        var apiConfig = configuration
-                .GetSection(nameof(AanHubApiConfiguration))
-                .Get<AanHubApiConfiguration>();
-
-        services.AddScoped<InnerApiAuthenticationHeaderHandler>();
+        var apiConfig = GetApiConfiguration(configuration, "AanHubApiConfiguration");
 
         services.AddRestEaseClient<IAanHubRestApiClient>(apiConfig.Url)
-            .AddHttpMessageHandler<InnerApiAuthenticationHeaderHandler>();
+            .AddHttpMessageHandler(() => new InnerApiAuthenticationHeaderHandler(new AzureClientCredentialHelper(), apiConfig.Identifier));
+
     }
 
     private static void AddCommitmentsV2ApiClient(IServiceCollection services, IConfiguration configuration)
     {
-        var apiConfig = configuration
-                .GetSection(nameof(ApprenticeAan.Configuration.CommitmentsV2ApiConfiguration))
-                .Get<ApprenticeAan.Configuration.CommitmentsV2ApiConfiguration>();
-
-        services.AddSingleton(apiConfig);
-
-        services.AddScoped<CommitmentsV2ApiHttpMessageHandler>();
+        var apiConfig = GetApiConfiguration(configuration, "CommitmentsV2ApiConfiguration");
 
         services.AddRestEaseClient<ICommitmentsV2InnerApiClient>(apiConfig.Url)
-            .AddHttpMessageHandler<CommitmentsV2ApiHttpMessageHandler>();
+            .AddHttpMessageHandler(() => new InnerApiAuthenticationHeaderHandler(new AzureClientCredentialHelper(), apiConfig.Identifier));
     }
+
+    private static void AddApprenticeAccountsApiClient(IServiceCollection services, IConfiguration configuration)
+    {
+        var apiConfig = GetApiConfiguration(configuration, "ApprenticeAccountsApiConfiguration");
+
+        services.AddRestEaseClient<IApprenticeAccountsApiClient>(apiConfig.Url)
+            .AddHttpMessageHandler(() => new InnerApiAuthenticationHeaderHandler(new AzureClientCredentialHelper(), apiConfig.Identifier));
+    }
+
+    private static void AddCoursesApiClient(IServiceCollection services, IConfiguration configuration)
+    {
+        var apiConfig = GetApiConfiguration(configuration, "CoursesApiConfiguration");
+
+        services.AddRestEaseClient<ICoursesApiClient>(apiConfig.Url)
+            .AddHttpMessageHandler(() => new InnerApiAuthenticationHeaderHandler(new AzureClientCredentialHelper(), apiConfig.Identifier));
+    }
+
+    private static void AddLocationApiClient(IServiceCollection services, IConfiguration configuration)
+    {
+        var apiConfig = GetApiConfiguration(configuration, "LocationApiConfiguration");
+
+        services.AddRestEaseClient<ILocationApiClient>(apiConfig.Url)
+            .AddHttpMessageHandler(() => new InnerApiAuthenticationHeaderHandler(new AzureClientCredentialHelper(), apiConfig.Identifier));
+    }
+
 
     private static void AddConfigurationOptions(IServiceCollection services, IConfiguration configuration)
     {
@@ -90,4 +110,7 @@ public static class ServiceCollectionExtensions
         services.Configure<CoursesApiConfiguration>(configuration.GetSection(nameof(CoursesApiConfiguration)));
         services.AddSingleton(c => c.GetService<IOptions<CoursesApiConfiguration>>()!.Value);
     }
+
+    private static InnerApiConfiguration GetApiConfiguration(IConfiguration configuration, string configurationName)
+        => configuration.GetSection(configurationName).Get<InnerApiConfiguration>()!;
 }
