@@ -3,59 +3,78 @@ using AutoFixture.NUnit3;
 using FluentAssertions;
 using Moq;
 using SFA.DAS.ApprenticeAan.Application.ApprenticeAccount.Queries.GetApprenticeAccount;
-using SFA.DAS.SharedOuterApi.Configuration;
-using SFA.DAS.SharedOuterApi.InnerApi.Requests;
-using SFA.DAS.SharedOuterApi.Interfaces;
-using SFA.DAS.SharedOuterApi.Models;
+using SFA.DAS.ApprenticeAan.Application.Infrastructure;
 using SFA.DAS.Testing.AutoFixture;
 
 namespace SFA.DAS.ApprenticeAan.Application.UnitTests.ApprenticeAccount.Queries.GetApprenticeAccount;
 
 public class GetApprenticeAccountQueryHandlerTests
 {
-    [Test]
-    [MoqAutoData]
+    [Test, MoqAutoData]
     public async Task Handle_InvokesApiClient(
-    [Frozen] Mock<IApprenticeAccountsApiClient<ApprenticeAccountsApiConfiguration>> apiClientMock,
+    [Frozen] Mock<IApprenticeAccountsApiClient> apiClientMock,
     GetApprenticeAccountQueryHandler sut,
-    GetApprenticeAccountQuery request,
+    Guid apprenticeId,
     GetApprenticeAccountQueryResult expectedResult,
     CancellationToken cancellationToken)
     {
+        const HttpStatusCode status = HttpStatusCode.OK;
 
-        apiClientMock.Setup(c => c.GetWithResponseCode<GetApprenticeAccountQueryResult>(It.Is<GetApprenticeRequest>(r => r.Id == request.ApprenticeId))).ReturnsAsync(new ApiResponse<GetApprenticeAccountQueryResult>(expectedResult, HttpStatusCode.OK, null));
+        var query = new GetApprenticeAccountQuery { ApprenticeId = apprenticeId };
+        apiClientMock.Setup(c => c.GetApprentice(apprenticeId, It.IsAny<CancellationToken>())).ReturnsAsync(
+                new RestEase.Response<GetApprenticeAccountQueryResult>(
+                    "not used",
+                    new HttpResponseMessage(status),
+                    () => expectedResult)
+            );
 
-        await sut.Handle(request, cancellationToken);
+        await sut.Handle(query, cancellationToken);
 
-        apiClientMock.Verify(c => c.GetWithResponseCode<GetApprenticeAccountQueryResult>(It.Is<GetApprenticeRequest>(r => r.Id == request.ApprenticeId)));
+        apiClientMock.Verify(c => c.GetApprentice(apprenticeId, It.IsAny<CancellationToken>()));
     }
 
-    [Test]
-    [MoqAutoData]
+    [Test, MoqAutoData]
     public async Task Handle_ApprenticeFound_ReturnsApprentice(
-        [Frozen] Mock<IApprenticeAccountsApiClient<ApprenticeAccountsApiConfiguration>> apiClientMock,
+        [Frozen] Mock<IApprenticeAccountsApiClient> apiClientMock,
         GetApprenticeAccountQueryHandler sut,
-        GetApprenticeAccountQuery request,
+        Guid apprenticeId,
         GetApprenticeAccountQueryResult expectedResult,
         CancellationToken cancellationToken)
     {
+        const HttpStatusCode status = HttpStatusCode.OK;
 
-        apiClientMock.Setup(c => c.GetWithResponseCode<GetApprenticeAccountQueryResult>(It.Is<GetApprenticeRequest>(r => r.Id == request.ApprenticeId))).ReturnsAsync(new ApiResponse<GetApprenticeAccountQueryResult>(expectedResult, HttpStatusCode.OK, null));
+        GetApprenticeAccountQuery request = new() { ApprenticeId = apprenticeId };
+
+        apiClientMock.Setup(c => c.GetApprentice(apprenticeId, It.IsAny<CancellationToken>())).ReturnsAsync(
+            new RestEase.Response<GetApprenticeAccountQueryResult>(
+                "not used",
+                new HttpResponseMessage(status),
+                () => expectedResult)
+        );
 
         var actualResult = await sut.Handle(request, cancellationToken);
 
         Assert.That(expectedResult, Is.EqualTo(actualResult));
     }
 
-    [Test]
-    [MoqAutoData]
-    public async Task Handle_ApprenticeNotFound_ReturnsApprentice(
-        [Frozen] Mock<IApprenticeAccountsApiClient<ApprenticeAccountsApiConfiguration>> apiClientMock,
+    [Test, MoqAutoData]
+    public async Task Handle_ApprenticeNotFound_ReturnsNullResult(
+        [Frozen] Mock<IApprenticeAccountsApiClient> apiClientMock,
         GetApprenticeAccountQueryHandler sut,
-        GetApprenticeAccountQuery request,
+        Guid apprenticeId,
         CancellationToken cancellationToken)
     {
-        apiClientMock.Setup(c => c.GetWithResponseCode<GetApprenticeAccountQueryResult?>(It.Is<GetApprenticeRequest>(r => r.Id == request.ApprenticeId))).ReturnsAsync(new ApiResponse<GetApprenticeAccountQueryResult?>(null, HttpStatusCode.NotFound, null));
+
+        const HttpStatusCode status = HttpStatusCode.NotFound;
+
+        GetApprenticeAccountQuery request = new() { ApprenticeId = apprenticeId };
+
+        apiClientMock.Setup(c => c.GetApprentice(apprenticeId, It.IsAny<CancellationToken>())).ReturnsAsync(
+            new RestEase.Response<GetApprenticeAccountQueryResult>(
+                "not used",
+                new HttpResponseMessage(status),
+                () => null!)
+        );
 
         var actualResult = await sut.Handle(request, cancellationToken);
 
@@ -65,12 +84,19 @@ public class GetApprenticeAccountQueryHandlerTests
     [Test]
     [MoqAutoData]
     public async Task Handle_ApiInvalidResponse_ThrowsInvalidOperationException(
-        [Frozen] Mock<IApprenticeAccountsApiClient<ApprenticeAccountsApiConfiguration>> apiClientMock,
+        [Frozen] Mock<IApprenticeAccountsApiClient> apiClientMock,
         GetApprenticeAccountQueryHandler sut,
         GetApprenticeAccountQuery request,
         CancellationToken cancellationToken)
     {
-        apiClientMock.Setup(c => c.GetWithResponseCode<GetApprenticeAccountQueryResult?>(It.Is<GetApprenticeRequest>(r => r.Id == request.ApprenticeId))).ReturnsAsync(new ApiResponse<GetApprenticeAccountQueryResult?>(null, HttpStatusCode.InternalServerError, null));
+        const HttpStatusCode status = HttpStatusCode.InternalServerError;
+
+        apiClientMock.Setup(c => c.GetApprentice(request.ApprenticeId, It.IsAny<CancellationToken>())).ReturnsAsync(
+            new RestEase.Response<GetApprenticeAccountQueryResult>(
+            "not used",
+            new HttpResponseMessage(status),
+            () => null!)
+            );
 
         Func<Task> act = () => sut.Handle(request, cancellationToken);
 
