@@ -1,5 +1,5 @@
 ﻿using MediatR;
-using SFA.DAS.ApprenticeAan.Infrastructure;
+using SFA.DAS.ApprenticeAan.Application.Infrastructure;
 
 namespace SFA.DAS.ApprenticeAan.Application.Employer.Queries.GetEmployerMemberSummary;
 
