@@ -19,7 +19,7 @@ using SFA.DAS.ApprenticeAan.Application.Regions.Queries.GetRegions;
 
 namespace SFA.DAS.ApprenticeAan.Application.Infrastructure;
 
-public interface IAanHubRestApiClient
+public interface IAanHubRestApiClient : IHealthChecker
 {
     [Get("/regions")]
     Task<GetRegionsQueryResult> GetRegions(CancellationToken cancellationToken);
