@@ -1,7 +1,6 @@
 ﻿using System.Diagnostics.CodeAnalysis;
 using System.Net.Http.Headers;
 using SFA.DAS.Api.Common.Interfaces;
-using SFA.DAS.ApprenticeAan.Api.Configuration;
 
 namespace SFA.DAS.ApprenticeAan.Application.Infrastructure;
 
@@ -9,20 +8,20 @@ namespace SFA.DAS.ApprenticeAan.Application.Infrastructure;
 public class InnerApiAuthenticationHeaderHandler : DelegatingHandler
 {
     private readonly IAzureClientCredentialHelper _azureClientCredentialHelper;
-    private readonly AanHubApiConfiguration _aanHubApiConfiguration;
+    private readonly string _apiIdentifier;
 
-    public InnerApiAuthenticationHeaderHandler(IAzureClientCredentialHelper azureClientCredentialHelper, AanHubApiConfiguration aanHubApiConfiguration)
+    public InnerApiAuthenticationHeaderHandler(IAzureClientCredentialHelper azureClientCredentialHelper, string apiIdentifier)
     {
         _azureClientCredentialHelper = azureClientCredentialHelper;
-        _aanHubApiConfiguration = aanHubApiConfiguration;
+        _apiIdentifier = apiIdentifier;
     }
 
     protected override async Task<HttpResponseMessage> SendAsync(HttpRequestMessage request, CancellationToken cancellationToken)
     {
         request.Headers.Add("X-Version", "1.0");
-        if (!string.IsNullOrEmpty(_aanHubApiConfiguration.Identifier))
+        if (!string.IsNullOrEmpty(_apiIdentifier))
         {
-            var accessToken = await _azureClientCredentialHelper.GetAccessTokenAsync(_aanHubApiConfiguration.Identifier);
+            var accessToken = await _azureClientCredentialHelper.GetAccessTokenAsync(_apiIdentifier);
             request.Headers.Authorization = new AuthenticationHeaderValue("Bearer", accessToken);
         }
         return await base.SendAsync(request, cancellationToken).ConfigureAwait(continueOnCapturedContext: false);
